unload plugins before moving files, and skip help.db unless upgrading the exe
Because a plugin might be holding a file open, we have to unload all the plugins
and then load them again when done. Also, because the help.db gets modified by
the help plugin, we need to only download it if the exe is changing otherwise
we'll always download it if the plugin has been used.