comm log updates
Added an optional flag to omit from log when using StoreFromOutside.
Intention on this is to have the ability to send info to the comm log
and not store it in the chat log saved to disk.

Added optional size limit so each comm log file saved to disk.
Value of 0 is no split, otherwise specified in MB.

The default value for add_timestamp in storeFromOutside is now true.