Powerup Trigger Updates (#327)
* Update soundpack's powerup trigger

Fix a minor soundpack issue caused by the reboot on 15 July
which included a change to the format of powerup notifications.

* Update buffer plugin's powerup trigger

Fix a minor review_buffers issue caused by the reboot on 15 July which
included a change to the format of powerup notifications. Also modify
the powerup trigger's body to save an abbreviated form of the message
for faster review.